Use from positional import positional syntax
This is always how i expected the class to be used and is why the WARN
and EXCEPT variables are on both the module and the class. It's much
neater and in the style of from pprint import pprint.

The tests should reflect the way we expect people to consume it.